this is the same script what in the 1.1 version

+++ caudium/branches/stable_1_0/debian/init.d
@@ -14,36 +14,45 @@
 # Caudium init.d startup file
 #
 # $Id$
-
+EXTVER=
 PATH=/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin
-DAEMON_DIR=/usr/lib/caudium
+DAEMON_DIR=/usr/lib/caudium${EXTVER}
 DAEMON=$DAEMON_DIR/start
-NAME=caudium
+NAME=caudium${EXTVER}
 DESC="Caudium Webserver"
 
-PIDFILE=/var/run/caudium/caudium.pid
+PIDFILE=/var/run/caudium${EXTVER}/caudium.pid
 DEFSTART_OPTIONS="--pid-file=$PIDFILE"
 
 test -f $DAEMON || exit 0
 
 set -e
 
-if test -f /etc/default/caudium; then
-    . /etc/default/caudium
+if test -f /etc/default/caudium${EXTVER}; then
+    . /etc/default/caudium${EXTVER}
 fi
 
 case "$1" in
   start)
+	if [ -f $PIDFILE ]
+	then
+		echo "PID file exists, Caudium already running ?"
+		/etc/init.d/caudium${EXTVER} stop
+		rm -f $PIDFILE
+		sleep 5
+	fi
 	echo -n "Starting $DESC: "
 	cd $DAEMON_DIR
 	$DAEMON $DEFSTART_OPTIONS $START_OPTIONS > /dev/null
 	echo "$NAME."
 	;;
   stop)
-	echo -n "Stopping $DESC: "
-	for p in `cat $PIDFILE`; do
-    	    kill -TERM $p > /dev/null || true
-	done
+    echo -n "Stopping $DESC: "
+    for p in `cat $PIDFILE`; do
+       if [ -n "`ps -p $p --no-headers`" ]; then
+            kill -TERM $p > /dev/null || true
+       fi
+    done
 	rm -f $PIDFILE
 	echo "$NAME."
 	;;
@@ -51,7 +60,9 @@ case "$1" in
 	if test -f $PIDFILE; then
 	    echo -n "Restarting $DESC: "
 	    for p in `cat $PIDFILE | sed -e 1d`; do
-		kill -HUP $p > /dev/null || true
+            if [ -n "`ps -p $p --no-headers`" ]; then
+                kill -HUP $p > /dev/null || true
+            fi
 	    done
 	    echo "$NAME."
 	else
